+++ privilege.go
@@ -61,6 +61,7 @@ func EnsureAdmin() {
 	exePath, err := os.Executable()
 	if err != nil {
 		fmt.Printf("%s\n", Tf("priv.exe_error", err))
+		RestoreConsoleMode()
 		os.Exit(1)
 	}
 
@@ -84,5 +85,6 @@ func EnsureAdmin() {
 	if ret <= 32 {
 		fmt.Printf("%s\n", Tf("priv.failed", ret))
 	}
+	RestoreConsoleMode()
 	os.Exit(0)
 }